backend/requirements.txt
Flask>=0.12.2
Flask-Babel>=0.11.2
Flask-HTTPAuth>=3.2.3
Flask-Login>=0.4.0
Flask-Mail>=0.9.1
Flask-Migrate>=2.1.1
Flask-SQLAlchemy>=2.3.2
idna>=2.6
itsdangerous>=0.24
Mako>=1.0.7
PyJWT>=1.5.3
python-dateutil>=2.6.1
python-dotenv>=0.7.1
python-editor>=1.0.3
pytz>=2017.2
redis>=2.10.6
requests>=2.18.4
rq>=0.9.2
six>=1.11.0
SQLAlchemy>=1.1.14
urllib3>=1.22
visitor>=0.1.3
Werkzeug>=0.12.2
rauth>=0.7.0
apispec>=0.39.0
marshmallow>=2.16.3
